Rename BookItem.module.css to HomePage.module.css
